parte 2 de dashboard de admin
