+++ front/club-deportivo/package.json
@@ -18,6 +18,7 @@
     "axios": "^1.7.9",
     "club-deportivo": "file:",
     "dotenv": "^16.4.7",
+    "lodash": "^4.17.21",
     "lucide-react": "^0.474.0",
     "motion": "^12.0.6",
     "next": "^15.1.6",
@@ -31,6 +32,7 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
+    "@types/lodash": "^4.17.15",
     "@types/node": "^20.17.16",
     "@types/react": "^19",
     "@types/react-dom": "^19",
