+++ front/club-deportivo/src/app/admin/usuarios/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import UsersContainer from '@/components/InfoAdmin/UsersContainer';
+
+export default function UsersPage() {
+  return <UsersContainer />;
+}
