+++ front/club-deportivo/src/components/InfoAdmin/Sidebar.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { Home, Table, Users, CreditCard, Settings, User, LogIn } from 'lucide-react';
 
 export default function Sidebar() {
@@ -12,7 +14,7 @@ export default function Sidebar() {
           <Home className="h-5 w-5" />
           <span>Dashboard</span>
         </a>
-        <a href="/usuarios" className="flex items-center space-x-3 p-3 rounded-lg hover:bg-blue-600/20">
+        <a href="/admin/usuarios" className="flex items-center space-x-3 p-3 rounded-lg hover:bg-blue-600/20">
           <Users className="h-5 w-5" />
           <span>Usuarios</span>
         </a>
